reduce bloat in index by running all experimental features in one function
